Fix call reference to be associated with the D channel message was received
The problem is that once a call reference was associated with a particular D
channel, it always was. This created an issue with NFAS when the secondary D
channel became active as the messages were still being sent on the non-active
D channel.

(closes issue #14959)
Reported by: remiq
Patches: 
      bug14959.patch uploaded by jpeeler (license 325)
Tested by: remiq


git-svn-id: http://svn.asterisk.org/svn/libpri/branches/1.4@1120 4e257579-0214-4147-9015-27880e0d027f